♻️  divide cmd helper function
