+++ cmdhelper/matrix_test.go
@@ -1,4 +1,4 @@
-package cmd
+package cmdhelper
 
 import (
 	"fmt"
